% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootKNN.R
\name{bootKNN}
\alias{bootKNN}
\title{bootstrapKNN}
\usage{
bootKNN(X, len, n, neighbour)
}
\arguments{
\item{X}{imput data (matrix)}

\item{len}{length of the first vector}

\item{n}{number of bootstrap repetitions}

\item{neighbour}{the number of nearest neighbors i.e. determining the value of k}
}
\value{
p_value
}
\description{
Based on multiple sampling, it calculates kNN statistics and compares them with the kNN statistics
calculated from the original sample.
}
\examples{
X <- matrix(rnorm(15, mean=5, sd=sqrt(3)), 5, 5)
Y <- matrix(rnorm(15, mean=5, sd=sqrt(3)), 6, 5)
macierz <- rbind(X, Y)
o <- rep(0, nrow(macierz4))
dane <- data.frame(macierz4, o)
dane
w <- length(X)
library(FNN)
bootKNN(dane, w, 10, nrow(dane)-1)

}
\keyword{bootstrap}
\keyword{kNN,}
