\docType{package}
\name{bootstrap-package}
\alias{bootstrap-package}
\title{Simple Bootstrapping for Hierarchical Clustering}
\description{
  This package provides bootstrapping for hierarchical
  clustering.
}
\details{
  \tabular{ll}{ Package: \tab bootstrap \cr Version: \tab
  0.1\cr Date: \tab 2013-06-24\cr License: \tab GPL (>=
  3)\cr URL: \tab http://www.github.com/sgibb/bootstrap/
  \cr }
}
\author{
  Sebastian Gibb <\email{mail@sebastiangibb.de}>
}
\references{
  \url{http://www.github.com/sgibb/bootstrap/}
}
\keyword{package}

