% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap2.R
\name{bootstrap2}
\alias{bootstrap2}
\title{A bootstrap function}
\usage{
bootstrap2(X, n, ES)
}
\arguments{
\item{X}{data vector}

\item{n}{number of bootstrap repetitions}

\item{ES}{estimation function, np. mean, var, sd or function create by you}
}
\description{
This functions allows you to estimate the value of the estimator
by means of multiple draws with returning from a sample.
}
\keyword{bootstrap,}
\keyword{sample}
