\name{bootstrap.data}
\title{Data sets of package bootstrap}
\alias{cell}
\alias{cholost}
\alias{hormone}
\alias{law}
\alias{law82}
\alias{lutenhorm}
\alias{mouse.c}
\alias{mouse.t}
\alias{scor}
\alias{spatial}
\alias{stamp}
\description{
  See Efron and Tibshirani (1993) for details on these
  datasets.
}
\keyword{datasets}
