\name{bootstrap.internal}
\title{Internal functions of package bootstrap}
\alias{ctsub}
\alias{xinter}
\alias{yinter}
\description{
  Internal functions of package bootstrap, these are not to be called by
  the user.
}
\usage{
ctsub(x, y, z)
xinter(x, y, z, increasing = T)
yinter(x, y, z, increasing = T)
}
\keyword{misc}
