% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmBoot.R
\name{lmBoot}
\alias{lmBoot}
\title{Bootstrapping regression coefficients}
\usage{
lmBoot(model, B = 1000, res = FALSE)
}
\arguments{
\item{model}{A linear model fit, produced by lm.}

\item{B}{Number of bootstrap replicates (default is set to 1000).}

\item{res}{Should we resample residuals? Setting res to FALSE indicates resampling of rows.}
}
\value{
A matrix of bootstrapped regression coefficients
}
\description{
Bootstrapping of linear model fits (using lm). Bootstrapping can be done by either resampling rows of the original data frame or resampling residuals from the original model fit.
}
\examples{
## resampling rows
data(airquality)
attach(airquality)
set.seed(30)
lmodel <- lm(Ozone ~ Wind)
lboot <- lmBoot(lmodel, B = 1000)

## resampling residuals
data(airquality)
attach(airquality)
set.seed(30)
lmodel <- lm(Ozone ~ Wind)
lboot <- lmBoot(lmodel, B = 1000, res = TRUE)

}
\author{
Soumik Purkayastha, \email{soumikp@umich.edu}
}
