% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneCV.R
\name{oneCV}
\alias{oneCV}
\title{Bootstrapping coefficient of variation}
\usage{
oneCV(x, B)
}
\arguments{
\item{x}{A vector of observed values of a random variable to be used for bootstrap.}

\item{B}{Number of bootstrap replicates (default is set to 1000).}
}
\value{
A vector of the bootstrapped coefficient of variation.
}
\description{
Obtain bootstrapped values of coefficient of variation.
}
\examples{
set.seed(1234)
oneCV(rnorm(100, 1, 1), 100)

}
\seealso{
\code{\link{oneSD}}, \code{\link{oneMean}}
}
\author{
Soumik Purkayastha, \email{soumikp@umich.edu}
}
