% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneMean.R
\name{oneMean}
\alias{oneMean}
\title{Bootstrapping mean}
\usage{
oneMean(x, B = 1000)
}
\arguments{
\item{x}{A vector of observed values of a random variable to be used for bootstrap.}

\item{B}{Number of boostrap replicates (default is set to 1000).}
}
\value{
A vector of the boostrapped mean.
}
\description{
Obtain bootstrapped values of mean.
}
\examples{
set.seed(1234)
oneMean(rnorm(100), 100)

}
\seealso{
\code{\link{oneRawMoment}}, \code{\link{oneCV}}
}
\author{
Soumik Purkayastha, \email{soumikp@umich.edu}
}
