% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneQuantile.R
\name{oneQuantile}
\alias{oneQuantile}
\title{Bootstrapping quantiles}
\usage{
oneQuantile(x, prob = 0.5, B = 1000)
}
\arguments{
\item{x}{A vector of observed values of a random variable to be used for bootstrap.}

\item{prob}{Order of quantile to be bootstrapped (default is set to 0.5).}

\item{B}{Number of bootstrap replicates (default is set to 1000).}
}
\value{
A vector of the bootstrapped quantile.
}
\description{
Obtain bootstrapped values of quantiles.
}
\examples{
set.seed(1234)
oneQuantile(rnorm(100, 0, 1), prob = 0.5, 100)

}
\seealso{
\code{\link{oneMedian}}
}
\author{
Soumik Purkayastha, \email{soumikp@umich.edu}
}
