% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneRawMoment.R
\name{oneRawMoment}
\alias{oneRawMoment}
\title{Bootstrapping raw moments}
\usage{
oneRawMoment(x, order = 2, B = 1000)
}
\arguments{
\item{x}{A vector of observed values of a random variable to be used for bootstrap.}

\item{order}{Order of raw/central moment to be bootstrapped (default is set to 2).}

\item{B}{Number of boostrap replicates (default is set to 1000).}
}
\value{
A vector of the boostrapped raw moment.
}
\description{
Obtain bootstrapped values of raw moments.
}
\examples{
set.seed(1234)
oneRawMoment(rnorm(100), order = 2, 100)

}
\seealso{
\code{\link{oneMean}}, \code{\link{oneCentralMoment}}
}
\author{
Soumik Purkayastha, \email{soumikp@umich.edu}
}
