% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneSD.R
\name{oneSD}
\alias{oneSD}
\title{Bootstrapping standard deviation}
\usage{
oneSD(x, B = 1000)
}
\arguments{
\item{x}{A vector of observed values of a random variable to be used for bootstrap.}

\item{B}{Number of bootstrap replicates (default is set to 1000).}
}
\value{
A vector of the bootstrapped standard deviation.
}
\description{
Obtain bootstrapped values of standard deviation.
}
\examples{
set.seed(1234)
oneSD(rnorm(100, 5, 2), 100)

}
\seealso{
\code{\link{oneCentralMoment}}, \code{\link{oneVar}}, \code{\link{oneCV}}
}
\author{
Soumik Purkayastha, \email{soumikp@umich.edu}
}
