% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneVar.R
\name{oneVar}
\alias{oneVar}
\title{Bootstrapping variance}
\usage{
oneVar(x, B = 1000)
}
\arguments{
\item{x}{A vector of observed values of a random variable to be used for bootstrap.}

\item{B}{Number of boostrap replicates (default is set to 1000).}
}
\value{
A vector of the boostrapped variance.
}
\description{
Obtain bootstrapped values of variance.
}
\examples{
set.seed(1234)
oneVar(rnorm(100, 5, 2), 100)

}
\seealso{
\code{\link{oneCentralMoment}}, \code{\link{oneSD}}
}
\author{
Soumik Purkayastha, \email{soumikp@umich.edu}
}
