% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoCor.R
\name{twoCor}
\alias{twoCor}
\title{Bootstrapping Pearson's correlation coefficient}
\usage{
twoCor(x, y, B = 1000)
}
\arguments{
\item{x}{One vector of observed values of a random variable to be used for bootstrap.}

\item{y}{Other vector of observed values of a random variable to be used for bootstrap (x and y must be of equal length).}

\item{B}{Number of bootstrap replicates (default is set to 1000).}
}
\value{
A vector of the bootstrapped Pearson's correlation coefficient.
}
\description{
Obtain bootstrapped values of Pearson's correlation coefficient.
}
\examples{
set.seed(1234)
x <- rnorm(100)
y <- 0.5*x + rnorm(100, 0.5)
twoCor(x, y, 100)

}
\seealso{
\code{\link{twoMeanDiff}}, \code{\link{twoQuantileDiff}}, \code{\link{twoVarRatio}}
}
\author{
Soumik Purkayastha, \email{soumikp@umich.edu}
}
