% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoMeanDiff.R
\name{twoMeanDiff}
\alias{twoMeanDiff}
\title{Bootstrapping difference of means}
\usage{
twoMeanDiff(x, y, B = 1000)
}
\arguments{
\item{x}{One vector of observed values of a random variable to be used for bootstrap.}

\item{y}{Other vector of observed values of a random variable to be used for bootstrap (x and y need not be of equal length).}

\item{B}{Number of bootstrap replicates (default is set to 1000).}
}
\value{
A vector of the bootstrapped difference of means.
}
\description{
Obtain bootstrapped values of difference of means.
}
\examples{
set.seed(1234)
twoMeanDiff(rnorm(100, 1, 1), rnorm(100, 5, 1), 100)

}
\seealso{
\code{\link{twoQuantileDiff}}, \code{\link{twoVarRatio}}, \code{\link{twoCor}}
}
\author{
Soumik Purkayastha, \email{soumikp@umich.edu}
}
