% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoQuantileDiff.R
\name{twoQuantileDiff}
\alias{twoQuantileDiff}
\title{Bootstrapping difference of quantiles}
\usage{
twoQuantileDiff(x, y, prob = 0.5, B = 1000)
}
\arguments{
\item{x}{One vector of observed values of a random variable to be used for bootstrap.}

\item{y}{Other vector of observed values of a random variable to be used for bootstrap (x and y need not be of equal length).}

\item{prob}{Order of quantile to be bootstrapped (default is set to 0.5).}

\item{B}{Number of bootstrap replicates (default is set to 1000).}
}
\value{
A vector of the bootstrapped difference of quantiles.
}
\description{
Obtain bootstrapped values of difference of quantiles.
}
\examples{
#this is used to bootstrap difference of medians
set.seed(1234)
twoQuantileDiff(rnorm(100, 1, 1), rnorm(100, 5, 1), 0.5, 100)

}
\seealso{
\code{\link{twoMeanDiff}}, \code{\link{twoVarRatio}}, \code{\link{twoCor}}
}
\author{
Soumik Purkayastha, \email{soumikp@umich.edu}
}
