% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoVarRatio.R
\name{twoVarRatio}
\alias{twoVarRatio}
\title{Bootstrapping ratio of variances}
\usage{
twoVarRatio(x, y, B = 1000)
}
\arguments{
\item{x}{One vector of observed values of a random variable to be used for bootstrap.}

\item{y}{Other vector of observed values of a random variable to be used for bootstrap (x and y need not be of equal length).}

\item{B}{Number of bootstrap replicates (default is set to 1000).}
}
\value{
A vector of the bootstrapped ratio of variances.
}
\description{
Obtain bootstrapped values of ratio of variances.
}
\examples{
set.seed(1234)
twoVarRatio(rnorm(100, 0, 1), rnorm(100, 0, 0.5), 100)

}
\seealso{
\code{\link{twoMeanDiff}}, \code{\link{twoQuantileDiff}}, \code{\link{twoCor}}
}
\author{
Soumik Purkayastha, \email{soumikp@umich.edu}
}
