% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-package.R
\docType{package}
\name{bootstrap-package}
\alias{bootstrap}
\alias{bootstrap-package}
\title{bootstrap: Bootstrap and Jackknife method}
\description{
This package contains a set of functions that used to estimate using multiple sample draws
}
\keyword{internal}
