% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-package.R
\docType{package}
\name{bootstrap-package}
\alias{bootstrap}
\alias{bootstrap-package}
\title{bootstrap: Bootstrap and Jackknife method}
\description{
This package contains a set of functions that used to estimate using multiple sample draws
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JuliaWieczorek/bootstrap}
  \item Report bugs at \url{https://github.com/JuliaWieczorek/bootstrap/issues}
}

}
\keyword{internal}
