% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{A bootstrap function}
\usage{
bootstrap(X, n, ES)
}
\arguments{
\item{X}{data vector, np. c(1,2,3)}

\item{n}{number of bootstrap repetitions}

\item{ES}{estimation function, np. mean, var, sd or function create by you}
}
\description{
This functions allows you to estimate the value of the estimator
by means of multiple draws with returning from a sample.
}
\keyword{bootstrap}
