\name{bootstrap-internal}
\alias{ctsub}
\alias{xinter}
\alias{yinter}
\title{Internal functions of package bootstrap}
\description{
  Internal functions of package bootstrap.
}
\usage{
ctsub(x, y, z)
xinter(x, y, z, increasing = TRUE)
yinter(x, y, z, increasing = TRUE)
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
