% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{jackknife}
\alias{jackknife}
\title{A jackknife function}
\usage{
jackknife(X, ES)
}
\arguments{
\item{X}{data vector}

\item{ES}{estimation function, np. mean, var, sd, median or function created by you}
}
\description{
The jackknife is a method used to estimate the variance
and bias of a large population.
This was the earliest resampling method, even before bootstrap method.
It involves a leave-one-out strategy of the estimation of a parameter
(e.g., the mean) in a data set of N observations (or records).
}
\keyword{bootstrap}
\keyword{jackknife,}
